+++ packages/opentelemetry-instrumentation-cohere/pyproject.toml
@@ -33,6 +33,14 @@ flake8 = "7.0.0"
 pytest = "7.4.4"
 pytest-sugar = "0.9.7"
 
+[tool.poetry.group.test.dependencies]
+cohere = "^4.42"
+pytest = "7.4.4"
+pytest-sugar = "0.9.7"
+vcrpy = "^6.0.1"
+pytest-recording = "^0.13.1"
+opentelemetry-sdk = "^1.22.0"
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
