+++ packages/opentelemetry-instrumentation-cohere/tests/cassettes/test_rerank/test_cohere_rerank.yaml
@@ -1,0 +1,69 @@
+interactions:
+- request:
+    body: '{"query": "What is the capital of the United States?", "documents": [{"text":
+      "Carson City is the capital city of the American state of Nevada. At the  2010
+      United States Census, Carson City had a population of 55,274."}, {"text": "The
+      Commonwealth of the Northern Mariana Islands is a group of islands in the Pacific
+      Ocean that are a political division controlled by the United States. Its capital
+      is Saipan."}, {"text": "Charlotte Amalie is the capital and largest city of
+      the United States Virgin Islands. It has about 20,000 people. The city is on
+      the island of Saint Thomas."}, {"text": "Washington, D.C. (also known as simply
+      Washington or D.C., and officially as the District of Columbia) is the capital
+      of the United States. It is a federal district. "}, {"text": "Capital punishment
+      (the death penalty) has existed in the United States since before the United
+      States was a country. As of 2017, capital punishment is legal in 30 of the 50
+      states."}, {"text": "North Dakota is a state in the United States. 672,591 people
+      lived in North Dakota in the year 2010. The capital and seat of government is
+      Bismarck."}], "model": "rerank-multilingual-v2.0", "top_n": 3, "return_documents":
+      false, "max_chunks_per_doc": null}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate, br
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '1224'
+      Content-Type:
+      - application/json
+      Request-Source:
+      - python-sdk-4.46
+      User-Agent:
+      - python-requests/2.31.0
+    method: POST
+    uri: https://api.cohere.ai/v1/rerank
+  response:
+    body:
+      string: '{"id":"de1152f3-d39e-42f2-832e-2bdd31278a09","results":[{"index":3,"relevance_score":0.9999989},{"index":2,"relevance_score":0.9003547},{"index":0,"relevance_score":0.6807725}],"meta":{"api_version":{"version":"1"},"billed_units":{"search_units":1}}}'
+    headers:
+      Alt-Svc:
+      - h3=":443"; ma=2592000,h3-29=":443"; ma=2592000
+      Content-Length:
+      - '250'
+      Via:
+      - 1.1 google
+      access-control-expose-headers:
+      - X-Debug-Trace-ID
+      content-type:
+      - application/json
+      date:
+      - Mon, 12 Feb 2024 15:03:51 GMT
+      server:
+      - envoy
+      vary:
+      - Origin
+      x-debug-trace-id:
+      - 46c68edf8dbd1ff1b55c15df703d7429
+      x-endpoint-monthly-call-limit:
+      - '500'
+      x-envoy-upstream-service-time:
+      - '29'
+      x-trial-endpoint-call-limit:
+      - '10'
+      x-trial-endpoint-call-remaining:
+      - '8'
+    status:
+      code: 200
+      message: OK
+version: 1
