--- packages/opentelemetry-instrumentation-cohere/tests/test_placeholder.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-    pass
